Version 1.3.1- pickling
I already wrote it, and maybe it runs better on someone else's machine.
This version of front will check if you have a saved version of the distributions before learning. This file, or its use, is not necessary! Figured I'd throw it in in case someone liked to burn through space on their solid state drive.